+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.6.0] - 2016-08-21
 ### Added
 - Add loading screen messages
 - Show Hero Power and Weapon details on mouse over
@@ -94,9 +96,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Code for initial development release
 
-[Unreleased]: https://github.com/HearthSim/joust/compare/0.5.0...HEAD
+[Unreleased]: https://github.com/HearthSim/joust/compare/0.6.0...HEAD
 [0.2.0]: https://github.com/HearthSim/joust/compare/0.1.0...0.2.0
 [0.3.0]: https://github.com/HearthSim/joust/compare/0.2.0...0.3.0
 [0.3.1]: https://github.com/HearthSim/joust/compare/0.3.0...0.3.1
 [0.4.0]: https://github.com/HearthSim/joust/compare/0.3.1...0.4.0
 [0.5.0]: https://github.com/HearthSim/joust/compare/0.4.0...0.5.0
+[0.6.0]: https://github.com/HearthSim/joust/compare/0.5.0...0.6.0